Style chart in dark mode
